# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Diem Tua Hoc Duong API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatMessageInput {
  sessionId: string;
  text: string;
}

export type ChatMessageSender = typeof ChatMessageSender[keyof typeof ChatMessageSender];


export const ChatMessageSender = {
  user: 'user',
  bot: 'bot',
} as const;

export interface ChatMessage {
  id: number;
  sessionId: string;
  text: string;
  sender: ChatMessageSender;
  createdAt: string;
}

export interface ChatResponse {
  userMessage: ChatMessage;
  botMessage: ChatMessage;
  suggestedReplies?: string[];
  /** @nullable */
  category?: string | null;
}

export interface Resource {
  id: number;
  title: string;
  summary: string;
  /** @nullable */
  content?: string | null;
  category: string;
  tags?: string[];
  readMinutes: number;
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
  createdAt: string;
}

export type MoodInputMood = typeof MoodInputMood[keyof typeof MoodInputMood];


export const MoodInputMood = {
  very_sad: 'very_sad',
  sad: 'sad',
  neutral: 'neutral',
  happy: 'happy',
  very_happy: 'very_happy',
} as const;

export interface MoodInput {
  sessionId: string;
  mood: MoodInputMood;
  score: number;
  /** @nullable */
  note?: string | null;
}

export interface MoodEntry {
  id: number;
  sessionId: string;
  mood: string;
  score: number;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface Hotline {
  id: number;
  name: string;
  phone: string;
  description: string;
  available24h: boolean;
  category?: string;
}

export interface FaqItem {
  id: number;
  question: string;
  answer: string;
}

export interface FaqCategory {
  id: number;
  name: string;
  questions: FaqItem[];
}

export interface StatsOverview {
  totalResources: number;
  totalHotlines: number;
  totalFaqItems: number;
  supportCategories: string[];
}

export type GetChatHistoryParams = {
sessionId: string;
limit?: number;
};

export type ListResourcesParams = {
category?: string;
limit?: number;
};

export type GetMoodHistoryParams = {
sessionId: string;
days?: number;
};