# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Diem Tua Hoc Duong API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageSender } from './chatMessageSender';

export interface ChatMessage {
  id: number;
  sessionId: string;
  text: string;
  sender: ChatMessageSender;
  createdAt: string;
}