# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Diem Tua Hoc Duong API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatResponse {
  userMessage: ChatMessage;
  botMessage: ChatMessage;
  suggestedReplies?: string[];
  /** @nullable */
  category?: string | null;
}