# lib/api-zod/src/generated/types/faqCategory.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Diem Tua Hoc Duong API
 * OpenAPI spec version: 0.1.0
 */
import type { FaqItem } from './faqItem';

export interface FaqCategory {
  id: number;
  name: string;
  questions: FaqItem[];
}