# lib/api-zod/src/generated/types/moodEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Diem Tua Hoc Duong API
 * OpenAPI spec version: 0.1.0
 */

export interface MoodEntry {
  id: number;
  sessionId: string;
  mood: string;
  score: number;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}