# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Diem Tua Hoc Duong API
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  id: number;
  title: string;
  summary: string;
  /** @nullable */
  content?: string | null;
  category: string;
  tags?: string[];
  readMinutes: number;
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
  createdAt: string;
}