# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Diem Tua Hoc Duong API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsOverview {
  totalResources: number;
  totalHotlines: number;
  totalFaqItems: number;
  supportCategories: string[];
}